Update listings to account for `cargo new` template update
Requires https://github.com/rust-lang/cargo/pull/13939.

+++ listings/ch11-writing-automated-tests/listing-11-03/src/lib.rs
@@ -1,5 +1,5 @@
 // ANCHOR: here
-pub fn add(left: usize, right: usize) -> usize {
+pub fn add(left: u64, right: u64) -> u64 {
     left + right
 }
